Simplify capability registration
